added `__file__` builtin for external scripts
external scripts and those published with RpsAddin now support the
`__file__` builtin - the variable contains the full path to the source
file being executed. for scripts that are read from a dll (command
loader assembly / rpsaddin) the absolute path to the dll is used as the
base directory of the script path (e.g.
'C:\myfolder\myaddin.dll\myscript.py')